ai-backend/pyproject.toml
[project]
name = "voiceai"
version = "0.1.0"
description = "Voice to voice AI assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.9.0",
    "numpy",
    "accelerate",
    "sounddevice",
    "soundfile",
    "pillow",
    "mss",
    "qwen-omni-utils",
    "flash-attn",
    "transformers",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
torch = { index = "pytorch" }
transformers = { git = "https://github.com/huggingface/transformers" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
